align changes with rest of project structure
