--- data/dates.json
@@ -1,130 +1,0 @@
-{
-  "days": {
-    "el": {
-      "full": ["Δευτέρα", "Τρίτη", "Τετάρτη", "Πέμπτη", "Παρασκευή", "Σάββατο", "Κυριακή"],
-      "short": ["Δευ", "Τρί", "Τετ", "Πέμ", "Παρ", "Σάβ", "Κυρ"],
-      "min": ["Δ", "Τ", "Τ", "Π", "Π", "Σ", "Κ"]
-    },
-    "en": {
-      "full": ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"],
-      "short": ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"],
-      "min": ["M", "T", "W", "T", "F", "S", "S"]
-    }
-  },
-  "months": {
-    "el": {
-      "full": [
-        "Ιανουάριος",
-        "Φεβρουάριος",
-        "Μάρτιος",
-        "Απρίλιος",
-        "Μάιος",
-        "Ιούνιος",
-        "Ιούλιος",
-        "Αύγουστος",
-        "Σεπτέμβριος",
-        "Οκτώβριος",
-        "Νοέμβριος",
-        "Δεκέμβριος"
-      ],
-      "alternative": [
-        "Ιανουαρίου",
-        "Φεβρουαρίου",
-        "Μαρτίου",
-        "Απριλίου",
-        "Μαΐου",
-        "Ιουνίου",
-        "Ιουλίου",
-        "Αυγούστου",
-        "Σεπτεμβρίου",
-        "Οκτωβρίου",
-        "Νοεμβρίου",
-        "Δεκεμβρίου"
-      ],
-      "short": ["Ιαν", "Φεβ", "Μάρ", "Απρ", "Μάι", "Ιούν", "Ιούλ", "Αύγ", "Σεπ", "Οκτ", "Νοέ", "Δεκ"],
-      "min": ["Ι", "Φ", "Μ", "Α", "Μ", "Ι", "Ι", "Α", "Σ", "Ο", "Ν", "Δ"]
-    },
-    "en": {
-      "full": [
-        "January",
-        "February",
-        "March",
-        "April",
-        "May",
-        "June",
-        "July",
-        "August",
-        "September",
-        "October",
-        "November",
-        "December"
-      ],
-      "alternative": [
-        "January",
-        "February",
-        "March",
-        "April",
-        "May",
-        "June",
-        "July",
-        "August",
-        "September",
-        "October",
-        "November",
-        "December"
-      ],
-      "short": ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"],
-      "min": ["J", "F", "M", "A", "M", "J", "J", "A", "S", "O", "N", "D"]
-    }
-  },
-  "quarters": {
-    "el": {
-      "full": ["1ο τρίμηνο", "2ο τρίμηνο", "3ο τρίμηνο", "4ο τρίμηνο"],
-      "short": ["Τ1", "Τ2", "Τ3", "Τ4"]
-    },
-    "en": {
-      "full": ["1st quarter", "2nd quarter", "3rd quarter", "4th quarter"],
-      "short": ["Q1", "Q2", "Q3", "Q4"]
-    }
-  },
-  "eras": {
-    "el": {
-      "full": ["προ Χριστού", "μετά Χριστόν"],
-      "short": ["π.Χ.", "μ.Χ."]
-    },
-    "en": {
-      "full": ["Before Christ", "Anno Domini"],
-      "short": ["BC", "AD"]
-    }
-  },
-  "holidays": {
-    "el": [
-      { "date": "01-01", "name": "Πρωτοχρονιά", "moveable": false },
-      { "date": "01-06", "name": "Θεοφάνεια", "moveable": false },
-      { "date": "03-25", "name": "Ευαγγελισμός της Θεοτόκου", "moveable": false },
-      { "date": "05-01", "name": "Εργατική Πρωτομαγιά", "moveable": false },
-      { "date": "08-15", "name": "Κοίμηση της Θεοτόκου", "moveable": false },
-      { "date": "10-28", "name": "Ημέρα του Όχι", "moveable": false },
-      { "date": "12-25", "name": "Χριστούγεννα", "moveable": false },
-      { "date": "12-26", "name": "Επόμενη των Χριστουγέννων", "moveable": false },
-      { "date": "cleanMonday", "name": "Καθαρά Δευτέρα", "moveable": true },
-      { "date": "goodFriday", "name": "Μεγάλη Παρασκευή", "moveable": true },
-      { "date": "easterMonday", "name": "Δευτέρα του Πάσχα", "moveable": true },
-      { "date": "pentecost", "name": "Αγίου Πνεύματος", "moveable": true }
-    ],
-    "en": [
-      { "date": "01-01", "name": "New Year's Day", "moveable": false },
-      { "date": "01-06", "name": "Epiphany", "moveable": false },
-      { "date": "03-25", "name": "Annunciation", "moveable": false },
-      { "date": "05-01", "name": "Labour Day", "moveable": false },
-      { "date": "08-15", "name": "Assumption of Mary", "moveable": false },
-      { "date": "10-28", "name": "Ohi Day", "moveable": false },
-      { "date": "12-25", "name": "Christmas Day", "moveable": false },
-      { "date": "12-26", "name": "Boxing Day", "moveable": false },
-      { "date": "cleanMonday", "name": "Clean Monday", "moveable": true },
-      { "date": "goodFriday", "name": "Good Friday", "moveable": true },
-      { "date": "easterMonday", "name": "Easter Monday", "moveable": true },
-      { "date": "pentecost", "name": "Pentecost", "moveable": true }
-    ]
-  }
-}
